% File ~/GitHub/MTH142Rtutorials/man/list_of_tutorials.Rd
\name{list_of_tutorials}
\alias{list_of_tutorials}
\title{This is a list of the MTH 142 R tutorials}
\description{
The MTH142Rtutorial library has no functions. Below is a list of the tutorials and a short description of each one.

  - 1st_Rtutorial_getting_started

  - 2nd_Rtutorial_variable_types_bar_hist

  - 3rd_Rtutorial_boxplot_sample_stats

  - 4th_Rtutorial_importing_data

  - 5th_Rtutorial_graphing

  - 6th_Rtutorial_norm_binom_distributions

  - 7th_Rtutorial_Prop_CI

  - 10th_Rtutorial_t_test

  - 11th_Rtutorial_diff_or_means

  - 12th_Rtutorial_ANOVA

  - 13th_Rtutorial_remove_outliers

  - 14th_Rtutorial_linear_regression

}


\examples{
# Run the 12th learnr tutorial

learnr::run_tutorial(12th_Rtutorial_ANOVA,package="MTH142Rtutorials")
}
\keyword{MTH142Rtutorial}

